re-organize a bunch of files
