+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/tuning/AutomaticFeedforwardTuner.java
@@ -1,7 +1,7 @@
 package org.firstinspires.ftc.teamcode.opmodes.tuning;
 
-import static org.firstinspires.ftc.teamcode.robot.DriveConstants.MAX_RPM;
-import static org.firstinspires.ftc.teamcode.robot.DriveConstants.rpmToVelocity;
+import static org.firstinspires.ftc.teamcode.robot.subsystems.drivetrain.DriveConstants.MAX_RPM;
+import static org.firstinspires.ftc.teamcode.robot.subsystems.drivetrain.DriveConstants.rpmToVelocity;
 
 import com.acmerobotics.dashboard.FtcDashboard;
 import com.acmerobotics.dashboard.config.Config;
@@ -13,7 +13,7 @@ import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
 import org.firstinspires.ftc.robotcore.internal.system.Misc;
-import org.firstinspires.ftc.teamcode.robot.CheckmateDrive;
+import org.firstinspires.ftc.teamcode.robot.subsystems.Drivetrain;
 import org.firstinspires.ftc.teamcode.robot.util.LoggingUtil;
 import org.firstinspires.ftc.teamcode.robot.util.RegressionUtil;
 
@@ -43,7 +43,7 @@ public class AutomaticFeedforwardTuner extends LinearOpMode {
     public void runOpMode() throws InterruptedException {
         telemetry = new MultipleTelemetry(telemetry, FtcDashboard.getInstance().getTelemetry());
 
-        CheckmateDrive drive = new CheckmateDrive(hardwareMap);
+        Drivetrain drive = new Drivetrain(hardwareMap);
 
         NanoClock clock = NanoClock.system();
