+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/robot/subsystems/drivetrain/trajectorysequence/EmptySequenceException.java
@@ -1,0 +1,4 @@
+package org.firstinspires.ftc.teamcode.robot.subsystems.drivetrain.trajectorysequence;
+
+
+public class EmptySequenceException extends RuntimeException { }
