--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/robot/trajectorysequence/EmptySequenceException.java
@@ -1,4 +1,0 @@
-package org.firstinspires.ftc.teamcode.robot.trajectorysequence;
-
-
-public class EmptySequenceException extends RuntimeException { }
